+++ history.md
@@ -4,6 +4,7 @@ Notable changes of coc.nvim:
 
 ## 2025-03-05
 
+- Export method `workspace.fixWin32unixFilepath` for filepath convert.
 - Add commands `document.enableInlayHint` and `document.disableInlayHint`.
 - Refresh popup menu when completing incomplete sources.